+++ js/ui/data_grid/ui.data_grid.columns_controller.js
@@ -1,5 +1,5 @@
 import gridCore from './ui.data_grid.core';
-import { columnsControllerModule } from '../grid_core/ui.grid_core.columns_controller';
+import { columnsControllerModule } from '../grid_core/columns_controller/module';
 import { extend } from '../../core/utils/extend';
 
 gridCore.registerModule('columns', {
